+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+-   `forms.getForm()` function
+
 ## 0.2.2 (2018-09-25)
 
 ### Added